requirements.txt
# local package
-e .

# external requirements
click>=7.0,<7.1
Sphinx>=2.2,<2.3
coverage>=4.5,<4.6
black==19.3b0
pandas>=0.25,<0.26
numpy>=1.17,<1.18
scikit-learn>=0.21,<0.22
matplotlib>=3.1,<3.2
seaborn>=0.9,<1.0
jupyter
dill>=0.3,<0.4
flake8>=3.7,<3.8
awscli>=1.16,<1.17
python-dotenv>=0.5.1
bayesian-optimization>=1.0,<1.1
